added logs, moved customer repo to customer publisher
